Add files via upload
Solves the default setting of disabled tap to click on touchpad. Tested on Lubuntu 20.04.
-
tap-to-click-on.py should be stored into /usr/bin and made executable by 'sudo chmod +x  /usr/bin/tap-to-click-on.py
tap-to-click-on.py assumes your python 3 executable is in /usr/bin : amend it if different on your system
-
lubuntu-upg-notifier.sh is a patched version of the originally installed script checking for updates; this is done to hijack the script and make sure the tap-to-click-on.py scrip is executed at every boot close to boot end. The only modified line is where " && /usr/bin/tap-to-click-on.py" is added.